add permission to bottom controls
