+++ template/src/SDKAppWrapper.tsx
@@ -33,6 +33,13 @@ export interface AppBuilderSdkApiInterface {
     preference?: {
       disableShareTile: boolean;
       disableVideoProcessors: boolean;
+      disableAudio?: boolean;
+      disableVideo?: boolean;
+      disableScreenShare?: boolean;
+      disableSettings?: boolean;
+      disableParticipantsPanel?: boolean;
+      disableChat?: boolean;
+      disableInvite?: boolean;
     },
   ) => Promise<
     [
