+++ template/src/components/Controls.tsx
@@ -9,1131 +9,43 @@
  information visit https://appbuilder.agora.io. 
 *********************************************
 */
-import React, {
-  useState,
-  useContext,
-  useEffect,
-  useRef,
-  useReducer,
-} from 'react';
+import React from 'react';
 import {View, StyleSheet, useWindowDimensions} from 'react-native';
 import {
-  DispatchContext,
-  PropsContext,
-  ToggleState,
-  useLocalUid,
-} from '../../agora-rn-uikit';
-import LocalAudioMute from '../subComponents/LocalAudioMute';
-import LocalVideoMute from '../subComponents/LocalVideoMute';
-import Recording from '../subComponents/Recording';
-import LocalSwitchCamera from '../subComponents/LocalSwitchCamera';
-import ScreenshareButton from '../subComponents/screenshare/ScreenshareButton';
-import isMobileOrTablet from '../utils/isMobileOrTablet';
-import {ClientRoleType} from '../../agora-rn-uikit';
-import LiveStreamControls from './livestream/views/LiveStreamControls';
-import {
   BREAKPOINTS,
-  isWeb,
-  isWebInternal,
   CustomToolbarMerge,
   CustomToolbarSorting,
-  MergeMoreButtonFields,
-  CustomToolbarSort,
 } from '../utils/common';
 import {RoomInfoContextInterface, useRoomInfo} from './room-info/useRoomInfo';
-import LocalEndcall from '../subComponents/LocalEndCall';
-import LayoutIconButton from '../subComponents/LayoutIconButton';
-import CopyJoinInfo from '../subComponents/CopyJoinInfo';
-import IconButton from '../atoms/IconButton';
-import ActionMenu, {ActionMenuItem} from '../atoms/ActionMenu';
-import useLayoutsData from '../pages/video-call/useLayoutsData';
-import {
-  ChatType,
-  SidePanelType,
-  useChatUIControls,
-  useContent,
-  useLayout,
-  useRecording,
-  useSidePanel,
-  useSpeechToText,
-} from 'customization-api';
-import {useVideoCall} from './useVideoCall';
-import {useScreenshare} from '../subComponents/screenshare/useScreenshare';
-import LayoutIconDropdown from '../subComponents/LayoutIconDropdown';
+import {useContent, useSpeechToText} from 'customization-api';
 import {useCaption} from '../../src/subComponents/caption/useCaption';
-import LanguageSelectorPopup from '../../src/subComponents/caption/LanguageSelectorPopup';
-import useSTTAPI from '../../src/subComponents/caption/useSTTAPI';
-import {EventNames} from '../rtm-events';
-import events, {PersistanceLevel} from '../rtm-events-api';
 import Toast from '../../react-native-toast-message';
 import {getLanguageLabel} from '../../src/subComponents/caption/utils';
 import Toolbar from '../atoms/Toolbar';
-import ToolbarItem, {useToolbarProps} from '../atoms/ToolbarItem';
 import {
   ToolbarPresetProps,
   ToolbarItemHide,
   ToolbarItemLabel,
-  ToolbarMoreButtonDefaultFields,
-  ToolbarMoreButtonCustomFields,
 } from '../atoms/ToolbarPreset';
-
-import {whiteboardContext} from './whiteboard/WhiteboardConfigure';
-import {RoomPhase} from 'white-web-sdk';
-import {useNoiseSupression} from '../app-state/useNoiseSupression';
-
-import {useVB} from './virtual-background/useVB';
-import WhiteboardWrapper from './whiteboard/WhiteboardWrapper';
-import LocalEventEmitter, {
-  LocalEventsEnum,
-} from '../rtm-events-api/LocalEvents';
 import {useSetRoomInfo} from './room-info/useSetRoomInfo';
 import {useString} from '../utils/useString';
 import {
   sttSpokenLanguageToastHeading,
   sttSpokenLanguageToastSubHeading,
-  toolbarItemCaptionText,
-  toolbarItemChatText,
-  toolbarItemInviteText,
-  toolbarItemLayoutText,
-  toolbarItemMoreText,
-  toolbarItemNoiseCancellationText,
-  toolbarItemPeopleText,
-  toolbarItemRecordingText,
-  toolbarItemViewRecordingText,
-  toolbarItemSettingText,
-  toolbarItemShareText,
-  toolbarItemTranscriptText,
-  toolbarItemVirtualBackgroundText,
-  toolbarItemWhiteboardText,
 } from '../language/default-labels/videoCallScreenLabels';
-import {LogSource, logger} from '../logger/AppBuilderLogger';
-import {useModal} from '../utils/useModal';
-import ViewRecordingsModal from './recordings/ViewRecordingsModal';
 import {filterObject} from '../utils/index';
 import {useLanguage} from '../language/useLanguage';
-import RecordingDeletePopup from './recordings/RecordingDeletePopup';
-
-export const useToggleWhiteboard = () => {
-  const {
-    whiteboardActive,
-    joinWhiteboardRoom,
-    leaveWhiteboardRoom,
-    getWhiteboardUid,
-  } = useContext(whiteboardContext);
-  const {setCustomContent} = useContent();
-  const {setLayout} = useLayout();
-  const {dispatch} = useContext(DispatchContext);
-  return () => {
-    if ($config.ENABLE_WHITEBOARD) {
-      if (whiteboardActive) {
-        leaveWhiteboardRoom();
-        setCustomContent(getWhiteboardUid(), false);
-        setLayout('grid');
-        events.send(
-          EventNames.WHITEBOARD_ACTIVE,
-          JSON.stringify({status: false}),
-          PersistanceLevel.Session,
-        );
-      } else {
-        joinWhiteboardRoom();
-        setCustomContent(getWhiteboardUid(), WhiteboardWrapper, {}, true);
-        dispatch({
-          type: 'UserPin',
-          value: [getWhiteboardUid()],
-        });
-        setLayout('pinned');
-        events.send(
-          EventNames.WHITEBOARD_ACTIVE,
-          JSON.stringify({status: true}),
-          PersistanceLevel.Session,
-        );
-      }
-    }
-  };
-};
-
-export const WhiteboardListener = () => {
-  const {dispatch} = useContext(DispatchContext);
-  const {setCustomContent} = useContent();
-  const {currentLayout, setLayout} = useLayout();
-  const {
-    data: {isHost},
-    isWhiteBoardOn,
-  } = useRoomInfo();
-
-  React.useEffect(() => {
-    if (($config.ENABLE_WAITING_ROOM && !isHost) || $config.AUTO_CONNECT_RTM) {
-      if (isWhiteBoardOn) {
-        WhiteboardStartedCallBack();
-      } else {
-        WhiteboardStoppedCallBack();
-      }
-    }
-  }, [isWhiteBoardOn, isHost]);
-
-  const WhiteboardCallBack = ({status}) => {
-    if (status) {
-      WhiteboardStartedCallBack();
-    } else {
-      WhiteboardStoppedCallBack();
-    }
-  };
-
-  useEffect(() => {
-    if (
-      !$config.ENABLE_WAITING_ROOM ||
-      ($config.ENABLE_WAITING_ROOM && isHost)
-    ) {
-      LocalEventEmitter.on(
-        LocalEventsEnum.WHITEBOARD_ACTIVE_LOCAL,
-        WhiteboardCallBack,
-      );
-
-      return () => {
-        LocalEventEmitter.on(
-          LocalEventsEnum.WHITEBOARD_ACTIVE_LOCAL,
-          WhiteboardCallBack,
-        );
-      };
-    }
-  }, [isHost]);
-
-  //whiteboard start
-
-  const {
-    whiteboardActive,
-    joinWhiteboardRoom,
-    leaveWhiteboardRoom,
-    getWhiteboardUid,
-  } = useContext(whiteboardContext);
-
-  const WhiteboardStoppedCallBack = () => {
-    toggleWhiteboard(true, false);
-  };
-
-  const WhiteboardStartedCallBack = () => {
-    toggleWhiteboard(false, false);
-  };
-
-  useEffect(() => {
-    whiteboardActive && currentLayout !== 'pinned' && setLayout('pinned');
-  }, []);
-
-  const toggleWhiteboard = (
-    whiteboardActive: boolean,
-    triggerEvent: boolean,
-  ) => {
-    if ($config.ENABLE_WHITEBOARD) {
-      if (whiteboardActive) {
-        leaveWhiteboardRoom();
-        setCustomContent(getWhiteboardUid(), false);
-        setLayout('grid');
-        triggerEvent &&
-          events.send(
-            EventNames.WHITEBOARD_ACTIVE,
-            JSON.stringify({status: false}),
-            PersistanceLevel.Session,
-          );
-      } else {
-        joinWhiteboardRoom();
-        setCustomContent(getWhiteboardUid(), WhiteboardWrapper, {}, true);
-        dispatch({
-          type: 'UserPin',
-          value: [getWhiteboardUid()],
-        });
-        setLayout('pinned');
-        triggerEvent &&
-          events.send(
-            EventNames.WHITEBOARD_ACTIVE,
-            JSON.stringify({status: true}),
-            PersistanceLevel.Session,
-          );
-      }
-    }
-  };
-  return null;
-};
-
-const MoreButton = (props: {fields: ToolbarMoreButtonDefaultFields}) => {
-  //recording delete
-  const [isRecordingDeletePopupVisible, setRecordingDeletePopupVisible] =
-    React.useState<boolean>(false);
-  const [recordingIdToDelete, setRecordingIdToDelete] = useState(0);
-  //recording delete
-
-  const {label} = useToolbarProps();
-  const {data} = useRoomInfo();
-  const noiseCancellationLabel = useString(toolbarItemNoiseCancellationText)();
-  const whiteboardLabel = useString<boolean>(toolbarItemWhiteboardText);
-  const captionLabel = useString<boolean>(toolbarItemCaptionText);
-  const transcriptLabel = useString<boolean>(toolbarItemTranscriptText);
-  const settingsLabel = useString(toolbarItemSettingText)();
-  const screenShareButton = useString<boolean>(toolbarItemShareText);
-  const recordingButton = useString<boolean>(toolbarItemRecordingText);
-  const viewRecordingsLabel = useString<boolean>(
-    toolbarItemViewRecordingText,
-  )();
-  const moreButtonLabel = useString(toolbarItemMoreText)();
-  const virtualBackgroundLabel = useString(toolbarItemVirtualBackgroundText)();
-  const chatLabel = useString(toolbarItemChatText)();
-  const inviteLabel = useString(toolbarItemInviteText)();
-  const peopleLabel = useString(toolbarItemPeopleText)();
-  const layoutLabel = useString(toolbarItemLayoutText)();
-  const {dispatch} = useContext(DispatchContext);
-  const {rtcProps} = useContext(PropsContext);
-  const {setCustomContent} = useContent();
-  const [_, setActionMenuVisible] = React.useState(false);
-  const [isHovered, setIsHovered] = useState(false);
-  const [isHoveredOnModal, setIsHoveredOnModal] = useState(false);
-  const {
-    modalOpen: isVRModalOpen,
-    setModalOpen: setVRModalOpen,
-    toggle: toggleVRModal,
-  } = useModal();
-  const moreBtnRef = useRef(null);
-  const {width: globalWidth, height: globalHeight} = useWindowDimensions();
-  const layouts = useLayoutsData();
-  const {currentLayout, setLayout} = useLayout();
-  const layout = layouts.findIndex(item => item.name === currentLayout);
-  const {setSidePanel, sidePanel} = useSidePanel();
-  const {
-    isCaptionON,
-    setIsCaptionON,
-    language: prevLang,
-    isSTTActive,
-    setIsSTTActive,
-    isSTTError,
-  } = useCaption();
-
-  const isTranscriptON = sidePanel === SidePanelType.Transcript;
-
-  const [isLanguagePopupOpen, setLanguagePopup] =
-    React.useState<boolean>(false);
-  const isFirstTimePopupOpen = React.useRef(false);
-  const STT_clicked = React.useRef(null);
-
-  const {start, restart} = useSTTAPI();
-  const {
-    data: {isHost},
-  } = useRoomInfo();
-  const {setShowInvitePopup, setShowStopRecordingPopup, setShowLayoutOption} =
-    useVideoCall();
-  const {isScreenshareActive, startScreenshare, stopScreenshare} =
-    useScreenshare();
-  const {isRecordingActive, startRecording, inProgress, deleteRecording} =
-    useRecording();
-  const {setChatType} = useChatUIControls();
-  const actionMenuitems: ActionMenuItem[] = [];
-
-  const {isNoiseSupressionEnabled, setNoiseSupression} = useNoiseSupression();
-
-  //AINS
-  if ($config.ENABLE_NOISE_CANCELLATION) {
-    actionMenuitems.push({
-      componentName: 'noise-cancellation',
-      order: 0,
-      toggleStatus: isNoiseSupressionEnabled === ToggleState.enabled,
-      disabled:
-        isNoiseSupressionEnabled === ToggleState.disabling ||
-        isNoiseSupressionEnabled === ToggleState.enabling,
-      isBase64Icon: true,
-      //@ts-ignore
-      icon: 'noise-cancellation',
-      iconColor: $config.SECONDARY_ACTION_COLOR,
-      textColor: $config.FONT_COLOR,
-      title: noiseCancellationLabel,
-      //isNoiseSupressionEnabled === ToggleState.enabled
-      onPress: () => {
-        setActionMenuVisible(false);
-        setNoiseSupression(p => !p);
-      },
-    });
-  }
-  //AINS
-
-  //virtual background
-  const {isVBActive, setIsVBActive} = useVB();
-
-  const toggleVB = () => {
-    if (isVBActive) {
-      setSidePanel(SidePanelType.None);
-    } else {
-      setSidePanel(SidePanelType.VirtualBackground);
-    }
-    setIsVBActive(prev => !prev);
-  };
-  if ($config.ENABLE_VIRTUAL_BACKGROUND && !$config.AUDIO_ROOM) {
-    actionMenuitems.push({
-      componentName: 'virtual-background',
-      order: 1,
-      isBase64Icon: true,
-      //@ts-ignore
-      icon: 'vb',
-      iconColor: $config.SECONDARY_ACTION_COLOR,
-      textColor: $config.FONT_COLOR,
-      //title: `${isVBActive ? 'Hide' : 'Show'} Virtual Background`,
-      title: virtualBackgroundLabel,
-      onPress: () => {
-        setActionMenuVisible(false);
-        toggleVB();
-      },
-    });
-  }
-  //virtual background
-
-  //whiteboard start
-
-  const {
-    whiteboardRoomState,
-    whiteboardActive,
-    joinWhiteboardRoom,
-    leaveWhiteboardRoom,
-    getWhiteboardUid,
-    whiteboardStartedFirst,
-  } = useContext(whiteboardContext);
-
-  const WhiteboardStoppedCallBack = () => {
-    toggleWhiteboard(true, false);
-  };
-
-  const WhiteboardStartedCallBack = () => {
-    toggleWhiteboard(false, false);
-  };
-
-  useEffect(() => {
-    whiteboardActive && currentLayout !== 'pinned' && setLayout('pinned');
-  }, []);
-
-  const WhiteboardCallBack = ({status}) => {
-    if (status) {
-      WhiteboardStartedCallBack();
-    } else {
-      WhiteboardStoppedCallBack();
-    }
-  };
-
-  useEffect(() => {
-    LocalEventEmitter.on(
-      LocalEventsEnum.WHITEBOARD_ACTIVE_LOCAL,
-      WhiteboardCallBack,
-    );
-    return () => {
-      LocalEventEmitter.off(
-        LocalEventsEnum.WHITEBOARD_ACTIVE_LOCAL,
-        WhiteboardCallBack,
-      );
-    };
-  }, []);
-
-  const toggleWhiteboard = (
-    whiteboardActive: boolean,
-    triggerEvent: boolean,
-  ) => {
-    if ($config.ENABLE_WHITEBOARD) {
-      if (whiteboardActive) {
-        leaveWhiteboardRoom();
-        setCustomContent(getWhiteboardUid(), false);
-        setLayout('grid');
-        triggerEvent &&
-          events.send(
-            EventNames.WHITEBOARD_ACTIVE,
-            JSON.stringify({status: false}),
-            PersistanceLevel.Session,
-          );
-      } else {
-        joinWhiteboardRoom();
-        setCustomContent(getWhiteboardUid(), WhiteboardWrapper, {}, true);
-        dispatch({
-          type: 'UserPin',
-          value: [getWhiteboardUid()],
-        });
-        setLayout('pinned');
-        triggerEvent &&
-          events.send(
-            EventNames.WHITEBOARD_ACTIVE,
-            JSON.stringify({status: true}),
-            PersistanceLevel.Session,
-          );
-      }
-    }
-  };
-  const WhiteboardDisabled =
-    !isHost ||
-    whiteboardRoomState === RoomPhase.Connecting ||
-    whiteboardRoomState === RoomPhase.Disconnecting;
-
-  //Disable whiteboard button when backend sends error
-  const WhiteboardError =
-    data?.whiteboard?.error &&
-    (data?.whiteboard?.error?.code || data?.whiteboard?.error?.message)
-      ? true
-      : false;
-
-  //whiteboard ends
-
-  if (isHost && $config.ENABLE_WHITEBOARD && isWebInternal()) {
-    actionMenuitems.push({
-      componentName: 'whiteboard',
-      order: 2,
-      disabled: WhiteboardDisabled,
-      isBase64Icon: true,
-      //@ts-ignore
-      icon: 'whiteboard-new',
-      iconColor: $config.SECONDARY_ACTION_COLOR,
-      textColor: $config.FONT_COLOR,
-      title: whiteboardLabel(whiteboardActive),
-      onPress: () => {
-        if (WhiteboardError) {
-          setActionMenuVisible(false);
-          Toast.show({
-            leadingIconName: 'alert',
-            type: 'error',
-            text1: 'Failed to enable Whiteboard Service.',
-            text2: data?.whiteboard?.error?.message,
-            visibilityTime: 1000 * 10,
-            primaryBtn: null,
-            secondaryBtn: null,
-            leadingIcon: null,
-          });
-          logger.error(
-            LogSource.Internals,
-            'JOIN_MEETING',
-            'Failed to enable Whiteboard Service',
-            {
-              message: data?.whiteboard?.error?.message,
-              code: data?.whiteboard?.error?.code,
-            },
-          );
-        } else {
-          setActionMenuVisible(false);
-          toggleWhiteboard(whiteboardActive, true);
-        }
-      },
-    });
-  }
-
-  // host can see stt options and attendee can view only when stt is enabled by a host in the channel
-
-  if ($config.ENABLE_STT && $config.ENABLE_CAPTION) {
-    actionMenuitems.push({
-      componentName: 'caption',
-      order: 3,
-      icon: `${isCaptionON ? 'captions-off' : 'captions'}`,
-      iconColor: $config.SECONDARY_ACTION_COLOR,
-      textColor: $config.FONT_COLOR,
-      disabled: !(
-        $config.ENABLE_STT &&
-        $config.ENABLE_CAPTION &&
-        (isHost || (!isHost && isSTTActive))
-      ),
-      title: captionLabel(isCaptionON),
-      onPress: () => {
-        setActionMenuVisible(false);
-        STT_clicked.current = !isCaptionON ? 'caption' : null;
-        if (isSTTError) {
-          setIsCaptionON(prev => !prev);
-          return;
-        }
-        if (isSTTActive) {
-          setIsCaptionON(prev => !prev);
-          // is lang popup has been shown once for any user in meeting
-        } else {
-          isFirstTimePopupOpen.current = true;
-          setLanguagePopup(true);
-        }
-      },
-    });
-
-    if ($config.ENABLE_MEETING_TRANSCRIPT) {
-      actionMenuitems.push({
-        componentName: 'transcript',
-        order: 4,
-        icon: 'transcript',
-        iconColor: $config.SECONDARY_ACTION_COLOR,
-        textColor: $config.FONT_COLOR,
-        disabled: !(
-          $config.ENABLE_STT &&
-          $config.ENABLE_CAPTION &&
-          $config.ENABLE_MEETING_TRANSCRIPT &&
-          (isHost || (!isHost && isSTTActive))
-        ),
-        title: transcriptLabel(isTranscriptON),
-        onPress: () => {
-          setActionMenuVisible(false);
-          STT_clicked.current = !isTranscriptON ? 'transcript' : null;
-          if (isSTTError) {
-            !isTranscriptON
-              ? setSidePanel(SidePanelType.Transcript)
-              : setSidePanel(SidePanelType.None);
-            return;
-          }
-          if (isSTTActive) {
-            !isTranscriptON
-              ? setSidePanel(SidePanelType.Transcript)
-              : setSidePanel(SidePanelType.None);
-          } else {
-            isFirstTimePopupOpen.current = true;
-            setLanguagePopup(true);
-          }
-        },
-      });
-    }
-  }
-
-  // view recordings
-
-  if (isHost && $config.CLOUD_RECORDING && isWeb()) {
-    actionMenuitems.push({
-      componentName: 'view-recordings',
-      order: 5,
-      icon: 'play-circle',
-      iconColor: $config.SECONDARY_ACTION_COLOR,
-      textColor: $config.FONT_COLOR,
-      title: viewRecordingsLabel,
-      onPress: () => {
-        toggleVRModal();
-      },
-    });
-  }
-
-  actionMenuitems.push({
-    hide: w => {
-      return w >= BREAKPOINTS.lg ? true : false;
-    },
-    componentName: 'participant',
-    order: 6,
-    icon: 'participants',
-    iconColor: $config.SECONDARY_ACTION_COLOR,
-    textColor: $config.FONT_COLOR,
-    title: peopleLabel,
-    onPress: () => {
-      setActionMenuVisible(false);
-      setSidePanel(SidePanelType.Participants);
-    },
-  });
-
-  if ($config.CHAT) {
-    //disable chat button when BE sends error on chat
-    const ChatError =
-      data?.chat?.error &&
-      (data?.chat?.error?.code || data?.chat?.error?.message)
-        ? true
-        : false;
-    actionMenuitems.push({
-      hide: w => {
-        return w >= BREAKPOINTS.lg ? true : false;
-      },
-      componentName: 'chat',
-      order: 7,
-      icon: 'chat-nav',
-      iconColor: $config.SECONDARY_ACTION_COLOR,
-      textColor: $config.FONT_COLOR,
-      title: chatLabel,
-      onPress: () => {
-        if (ChatError) {
-          setActionMenuVisible(false);
-          Toast.show({
-            leadingIconName: 'alert',
-            type: 'error',
-            text1: 'Failed to enable Chat Service.',
-            text2: data?.chat?.error?.message,
-            visibilityTime: 1000 * 10,
-            primaryBtn: null,
-            secondaryBtn: null,
-            leadingIcon: null,
-          });
-          logger.error(
-            LogSource.Internals,
-            'JOIN_MEETING',
-            'Failed to enable Chat Service',
-            {
-              message: data?.chat?.error?.message,
-              code: data?.chat?.error?.code,
-            },
-          );
-        } else {
-          setActionMenuVisible(false);
-          setChatType(ChatType.Group);
-          setSidePanel(SidePanelType.Chat);
-        }
-      },
-    });
-  }
-
-  if ($config.SCREEN_SHARING) {
-    if (
-      !(
-        rtcProps.role == ClientRoleType.ClientRoleAudience &&
-        $config.EVENT_MODE &&
-        !$config.RAISE_HAND
-      )
-    ) {
-      actionMenuitems.push({
-        hide: w => {
-          return w >= BREAKPOINTS.sm ? true : false;
-        },
-        componentName: 'screenshare',
-        order: 8,
-        disabled:
-          rtcProps.role == ClientRoleType.ClientRoleAudience &&
-          $config.EVENT_MODE &&
-          $config.RAISE_HAND &&
-          !isHost,
-        icon: isScreenshareActive ? 'stop-screen-share' : 'screen-share',
-        iconColor: isScreenshareActive
-          ? $config.SEMANTIC_ERROR
-          : $config.SECONDARY_ACTION_COLOR,
-        textColor: isScreenshareActive
-          ? $config.SEMANTIC_ERROR
-          : $config.FONT_COLOR,
-        title: screenShareButton(isScreenshareActive),
-        onPress: () => {
-          setActionMenuVisible(false);
-          isScreenshareActive ? stopScreenshare() : startScreenshare();
-        },
-      });
-    }
-  }
-  if (isHost && $config.CLOUD_RECORDING) {
-    actionMenuitems.push({
-      hide: w => {
-        return w >= BREAKPOINTS.sm ? true : false;
-      },
-      componentName: 'recording',
-      order: 9,
-      disabled: inProgress,
-      icon: isRecordingActive ? 'stop-recording' : 'recording',
-      iconColor: isRecordingActive
-        ? $config.SEMANTIC_ERROR
-        : $config.SECONDARY_ACTION_COLOR,
-      textColor: isRecordingActive
-        ? $config.SEMANTIC_ERROR
-        : $config.FONT_COLOR,
-      title: recordingButton(isRecordingActive),
-      onPress: () => {
-        setActionMenuVisible(false);
-        if (!isRecordingActive) {
-          startRecording();
-        } else {
-          setShowStopRecordingPopup(true);
-        }
-      },
-    });
-  }
-
-  actionMenuitems.push({
-    hide: w => {
-      return w >= BREAKPOINTS.lg ? true : false;
-    },
-    componentName: 'layout',
-    order: 10,
-    //below icon key is dummy value
-    icon: 'grid',
-    externalIconString: layouts[layout]?.icon,
-    isExternalIcon: true,
-    iconColor: $config.SECONDARY_ACTION_COLOR,
-    textColor: $config.FONT_COLOR,
-    title: layoutLabel,
-    onPress: () => {
-      //setShowLayoutOption(true);
-    },
-    onHoverCallback: isHovered => {
-      setShowLayoutOption(isHovered);
-    },
-    onHoverContent: (
-      <LayoutIconDropdown
-        onHoverPlaceHolder="vertical"
-        setShowDropdown={() => {}}
-        showDropdown={true}
-        modalPosition={
-          globalWidth <= BREAKPOINTS.lg
-            ? {bottom: 65, left: -150}
-            : {bottom: 20, left: -150}
-        }
-        caretPosition={{bottom: 45, right: -10}}
-      />
-    ),
-  });
-
-  actionMenuitems.push({
-    hide: w => {
-      return w >= BREAKPOINTS.lg ? true : false;
-    },
-    componentName: 'invite',
-    order: 11,
-    icon: 'share',
-    iconColor: $config.SECONDARY_ACTION_COLOR,
-    textColor: $config.FONT_COLOR,
-    title: inviteLabel,
-    onPress: () => {
-      setActionMenuVisible(false);
-      setShowInvitePopup(true);
-    },
-  });
-
-  actionMenuitems.push({
-    hide: w => {
-      return w >= BREAKPOINTS.lg ? true : false;
-    },
-    componentName: 'settings',
-    order: 12,
-    icon: 'settings',
-    iconColor: $config.SECONDARY_ACTION_COLOR,
-    textColor: $config.FONT_COLOR,
-    title: settingsLabel,
-    onPress: () => {
-      setActionMenuVisible(false);
-      setSidePanel(SidePanelType.Settings);
-    },
-  });
-
-  useEffect(() => {
-    if (isHovered) {
-      setActionMenuVisible(true);
-    } else setActionMenuVisible(false);
-  }, [isHovered]);
-
-  useEffect(() => {
-    //hide action menu when user change layout
-    setActionMenuVisible(false);
-  }, [currentLayout]);
-
-  const onConfirm = async (langChanged, language) => {
-    const isCaptionClicked = STT_clicked.current === 'caption';
-    const isTranscriptClicked = STT_clicked.current === 'transcript';
-    setLanguagePopup(false);
-    isFirstTimePopupOpen.current = false;
-    const method = isCaptionClicked
-      ? isCaptionON
-      : isTranscriptON
-      ? 'stop'
-      : 'start';
-    if (isTranscriptClicked) {
-      if (!isTranscriptON) {
-        setSidePanel(SidePanelType.Transcript);
-      } else {
-        setSidePanel(SidePanelType.None);
-      }
-    }
-    if (method === 'stop') return; // not closing the stt service as it will stop for whole channel
-    if (method === 'start' && isSTTActive === true) return; // not triggering the start service if STT Service already started by anyone else in the channel
-
-    if (isCaptionClicked) {
-      setIsCaptionON(prev => !prev);
-    } else {
-    }
-
-    try {
-      const res = await start(language);
-      if (res?.message.includes('STARTED')) {
-        // channel is already started now restart
-        await restart(language);
-      }
-    } catch (error) {
-      logger.error(LogSource.Internals, 'STT', 'error in starting stt', error);
-    }
-  };
-
-  const {width, height} = useWindowDimensions();
-
-  const isHidden = (hide: ToolbarItemHide = false) => {
-    try {
-      return typeof hide === 'boolean'
-        ? hide
-        : typeof hide === 'function'
-        ? hide(width, height)
-        : false;
-    } catch (error) {
-      console.log('debugging isHidden error', error);
-      return false;
-    }
-  };
-
-  const moreButtonFields = props?.fields || {};
-
-  const ActionMenuItems = MergeMoreButtonFields(
-    actionMenuitems,
-    moreButtonFields,
-  )
-    ?.filter(i => !isHidden(i))
-    ?.sort(CustomToolbarSort);
-
-  const onRecordingDeleteConfirmation = () => {
-    setRecordingDeletePopupVisible(false);
-    deleteRecording(recordingIdToDelete)
-      .then(() => {
-        //To inform other user -> refresh the recording list
-        //in case recording list opened
-        events.send(
-          EventNames.RECORDING_DELETED,
-          JSON.stringify({recordingId: recordingIdToDelete}),
-          PersistanceLevel.None,
-        );
-        Toast.show({
-          leadingIconName: 'alert',
-          type: 'success',
-          text1: 'Recording has been deleted successfully.',
-          visibilityTime: 3000,
-          primaryBtn: null,
-          secondaryBtn: null,
-          leadingIcon: null,
-        });
-        setRecordingIdToDelete(0);
-        //to reopen recording list again
-        setTimeout(() => {
-          setVRModalOpen(true);
-        }, 3000);
-      })
-      .catch(() => {
-        Toast.show({
-          leadingIconName: 'alert',
-          type: 'error',
-          text1: 'Failed to delete the recording. Please try again later',
-          visibilityTime: 1000 * 10,
-          primaryBtn: null,
-          secondaryBtn: null,
-          leadingIcon: null,
-        });
-        setRecordingIdToDelete(0);
-      });
-  };
-
-  const onRecordingDeleteCancel = () => {
-    setRecordingIdToDelete(0);
-    toggleVRModal();
-  };
-
-  return (
-    <>
-      <LanguageSelectorPopup
-        modalVisible={isLanguagePopupOpen}
-        setModalVisible={setLanguagePopup}
-        onConfirm={onConfirm}
-        isFirstTimePopupOpen={isFirstTimePopupOpen.current}
-      />
-      {$config.CLOUD_RECORDING && isHost && isWeb() && (
-        <>
-          <RecordingDeletePopup
-            modalVisible={isRecordingDeletePopupVisible}
-            setModalVisible={setRecordingDeletePopupVisible}
-            onConfirm={onRecordingDeleteConfirmation}
-            onCancel={onRecordingDeleteCancel}
-          />
-          {isVRModalOpen ? (
-            <ViewRecordingsModal
-              setModalOpen={setVRModalOpen}
-              onDeleteAction={id => {
-                setRecordingIdToDelete(id);
-                toggleVRModal();
-                setRecordingDeletePopupVisible(true);
-              }}
-            />
-          ) : (
-            <></>
-          )}
-        </>
-      )}
-      <ActionMenu
-        containerStyle={globalWidth < 720 ? {width: 180} : {width: 260}}
-        hoverMode={true}
-        onHover={isVisible => setIsHoveredOnModal(isVisible)}
-        from={'control-bar'}
-        actionMenuVisible={isHovered || isHoveredOnModal}
-        setActionMenuVisible={setActionMenuVisible}
-        modalPosition={{
-          bottom: 8,
-          left: 0,
-        }}
-        items={ActionMenuItems}
-      />
-      <div
-        onMouseEnter={() => {
-          setIsHovered(true);
-        }}
-        onMouseLeave={() => {
-          setIsHovered(false);
-        }}>
-        {/** placeholder to hovering */}
-        <View
-          style={{
-            position: 'absolute',
-            top: -20,
-            zIndex: -1,
-            height: '50%',
-            width: '100%',
-            backgroundColor: 'transparent',
-          }}
-        />
-        <IconButton
-          setRef={ref => {
-            moreBtnRef.current = ref;
-          }}
-          onPress={() => {
-            //setActionMenuVisible(true);
-          }}
-          iconProps={{
-            name: 'more-menu',
-            tintColor: $config.SECONDARY_ACTION_COLOR,
-          }}
-          btnTextProps={{
-            text: $config.ICON_TEXT ? label || moreButtonLabel : '',
-            textColor: $config.FONT_COLOR,
-          }}
-        />
-      </div>
-    </>
-  );
-};
-export const LayoutToolbarItem = props => (
-  <ToolbarItem testID="layout-btn" collapsable={false} toolbarProps={props}>
-    {/**
-     * .measure returns undefined on Android unless collapsable=false or onLayout are specified
-     * so added collapsable property
-     * https://github.com/facebook/react-native/issues/29712
-     * */}
-    <LayoutIconButton />
-  </ToolbarItem>
-);
-export const InviteToolbarItem = props => {
-  return (
-    <ToolbarItem testID="invite-btn" toolbarProps={props}>
-      <CopyJoinInfo />
-    </ToolbarItem>
-  );
-};
-
-export const RaiseHandToolbarItem = props => {
-  const {rtcProps} = useContext(PropsContext);
-  // attendee can view option if any host has started STT
-  const {
-    data: {isHost},
-  } = useRoomInfo();
-  return $config.EVENT_MODE ? (
-    rtcProps.role == ClientRoleType.ClientRoleAudience ? (
-      <LiveStreamControls showControls={true} customProps={props} />
-    ) : rtcProps?.role == ClientRoleType.ClientRoleBroadcaster ? (
-      /**
-       * In event mode when raise hand feature is active
-       * and audience is promoted to host, the audience can also
-       * demote himself
-       */
-      <LiveStreamControls showControls={!isHost} customProps={props} />
-    ) : (
-      <></>
-    )
-  ) : (
-    <></>
-  );
-};
-
-export const LocalAudioToolbarItem = props => {
-  return (
-    <ToolbarItem testID="localAudio-btn" toolbarProps={props}>
-      <LocalAudioMute
-        showToolTip={true}
-        iconBGColor={props?.iconBGColor}
-        iconSize={props?.iconSize}
-        containerStyle={props?.containerStyle}
-      />
-    </ToolbarItem>
-  );
-};
-
-export const LocalVideoToolbarItem = props => {
-  return (
-    !$config.AUDIO_ROOM && (
-      <ToolbarItem testID="localVideo-btn" toolbarProps={props}>
-        <LocalVideoMute showToolTip={true} />
-      </ToolbarItem>
-    )
-  );
-};
-
-export const SwitchCameraToolbarItem = props => {
-  return (
-    !$config.AUDIO_ROOM &&
-    isMobileOrTablet() && (
-      <ToolbarItem testID="switchCamera-btn" toolbarProps={props}>
-        <LocalSwitchCamera />
-      </ToolbarItem>
-    )
-  );
-};
-
-export const ScreenShareToolbarItem = props => {
-  return (
-    $config.SCREEN_SHARING &&
-    !isMobileOrTablet() && (
-      <ToolbarItem testID="screenShare-btn" toolbarProps={props}>
-        <ScreenshareButton />
-      </ToolbarItem>
-    )
-  );
-};
-export const RecordingToolbarItem = props => {
-  const {
-    data: {isHost},
-  } = useRoomInfo();
-  return (
-    isHost &&
-    $config.CLOUD_RECORDING && (
-      <ToolbarItem testID="recording-btn" toolbarProps={props}>
-        <Recording />
-      </ToolbarItem>
-    )
-  );
-};
-
-export const MoreButtonToolbarItem = (props?: {
-  fields?: ToolbarMoreButtonCustomFields;
-}) => {
-  const {width} = useWindowDimensions();
-  const {
-    data: {isHost},
-  } = useRoomInfo();
-  const {isSTTActive} = useCaption();
-  const [_, forceUpdate] = useReducer(x => x + 1, 0);
-
-  useEffect(() => {
-    forceUpdate();
-  }, [isHost]);
-
-  return width < BREAKPOINTS.lg ||
-    ($config.ENABLE_STT &&
-      $config.ENABLE_CAPTION &&
-      (isHost || (!isHost && isSTTActive))) ||
-    $config.ENABLE_NOISE_CANCELLATION ||
-    (isHost && $config.CLOUD_RECORDING && isWeb()) ||
-    ($config.ENABLE_VIRTUAL_BACKGROUND && !$config.AUDIO_ROOM) ||
-    (isHost && $config.ENABLE_WHITEBOARD && isWebInternal()) ? (
-    <ToolbarItem testID="more-btn" toolbarProps={props}>
-      {((!$config.AUTO_CONNECT_RTM && !isHost) || $config.AUTO_CONNECT_RTM) &&
-      $config.ENABLE_WHITEBOARD &&
-      isWebInternal() ? (
-        <WhiteboardListener />
-      ) : (
-        <></>
-      )}
-      <MoreButton fields={props?.fields} />
-    </ToolbarItem>
-  ) : (
-    <WhiteboardListener />
-  );
-};
-export interface LocalEndcallToolbarItemProps {
-  customExit?: () => void;
-}
-export const LocalEndcallToolbarItem = (
-  props?: LocalEndcallToolbarItemProps,
-) => {
-  return (
-    <ToolbarItem
-      testID={props?.customExit ? 'endCall-btn-custom' : 'endCall-btn'}
-      toolbarProps={props}>
-      <LocalEndcall {...props} />
-    </ToolbarItem>
-  );
-};
+import {LayoutToolbarItem} from './controls/toolbar-items/LayoutToolbarItem';
+import {InviteToolbarItem} from './controls/toolbar-items/InviteToolbarItem';
+import {RaiseHandToolbarItem} from './controls/toolbar-items/RaiseHandToolbarItem';
+import {LocalAudioToolbarItem} from './controls/toolbar-items/LocalAudioToolbarItem';
+import {LocalVideoToolbarItem} from './controls/toolbar-items/LocalVideoToolbarItem';
+import {SwitchCameraToolbarItem} from './controls/toolbar-items/SwitchCameraToolbarItem';
+import {ScreenShareToolbarItem} from './controls/toolbar-items/ScreenShareToolbarItem';
+import {RecordingToolbarItem} from './controls/toolbar-items/RecordingToolbarItem';
+import {MoreButtonToolbarItem} from './controls/toolbar-items/MoreButtonToolbarItem';
+import {LocalEndcallToolbarItem} from './controls/toolbar-items/LocalEndcallToolbarItem';
+import {withControlPermissionHOC} from './controls/useControlPermissionMatrix';
 
 const defaultItems: ToolbarPresetProps['items'] = {
   layout: {
@@ -1146,7 +58,7 @@ const defaultItems: ToolbarPresetProps['items'] = {
   },
   invite: {
     align: 'start',
-    component: InviteToolbarItem,
+    component: withControlPermissionHOC(InviteToolbarItem, 'inviteControl'),
     order: 1,
     hide: w => {
       return w < BREAKPOINTS.lg ? true : false;
@@ -1174,7 +86,10 @@ const defaultItems: ToolbarPresetProps['items'] = {
   },
   screenshare: {
     align: 'center',
-    component: ScreenShareToolbarItem,
+    component: withControlPermissionHOC(
+      ScreenShareToolbarItem,
+      'screenshareControl',
+    ),
     order: 4,
     hide: w => {
       return w < BREAKPOINTS.sm ? true : false;
@@ -1204,6 +119,7 @@ export interface ControlsProps {
   items?: ToolbarPresetProps['items'];
   includeDefaultItems?: boolean;
 }
+
 const Controls = (props: ControlsProps) => {
   const {languageCode} = useLanguage();
   const {items = {}, includeDefaultItems = true} = props;
@@ -1313,12 +229,10 @@ const Controls = (props: ControlsProps) => {
       return false;
     }
   };
-
   const mergedItems = CustomToolbarMerge(
     includeDefaultItems ? defaultItems : {},
     items,
   );
-
   const startItems = filterObject(
     mergedItems,
     ([_, v]) => v?.align === 'start' && !isHidden(v?.hide),
