+++ template/src/components/SdkApiContext.tsx
@@ -30,6 +30,13 @@ type SdkApiContextInterface = {
         preference: {
           disableShareTile: boolean;
           disableVideoProcessors: boolean;
+          disableAudio?: boolean;
+          disableVideo?: boolean;
+          disableScreenShare?: boolean;
+          disableSettings?: boolean;
+          disableParticipantsPanel?: boolean;
+          disableChat?: boolean;
+          disableInvite?: boolean;
         };
         promise: extractPromises<_InternalSDKMethodEventsMap['join']>;
       }
