+++ template/src/components/controls/toolbar-items/LocalVideoToolbarItem.tsx
@@ -1,0 +1,13 @@
+import React from 'react';
+import ToolbarItem from '../../../atoms/ToolbarItem';
+import LocalVideoMute from '../../../subComponents/LocalVideoMute';
+
+export const LocalVideoToolbarItem = props => {
+  return (
+    !$config.AUDIO_ROOM && (
+      <ToolbarItem testID="localVideo-btn" toolbarProps={props}>
+        <LocalVideoMute showToolTip={true} />
+      </ToolbarItem>
+    )
+  );
+};
