+++ template/src/components/controls/toolbar-items/SwitchCameraToolbarItem.tsx
@@ -1,0 +1,15 @@
+import React from 'react';
+import ToolbarItem from '../../../atoms/ToolbarItem';
+import isMobileOrTablet from '../../../utils/isMobileOrTablet';
+import LocalSwitchCamera from '../../../subComponents/LocalSwitchCamera';
+
+export const SwitchCameraToolbarItem = props => {
+  return (
+    !$config.AUDIO_ROOM &&
+    isMobileOrTablet() && (
+      <ToolbarItem testID="switchCamera-btn" toolbarProps={props}>
+        <LocalSwitchCamera />
+      </ToolbarItem>
+    )
+  );
+};
