+++ template/src/utils/SdkMethodEvents.ts
@@ -14,6 +14,13 @@ export interface SdkMethodEvents {
     preference?: {
       disableShareTile: boolean;
       disableVideoProcessors: boolean;
+      disableAudio?: boolean;
+      disableVideo?: boolean;
+      disableScreenShare?: boolean;
+      disableSettings?: boolean;
+      disableParticipantsPanel?: boolean;
+      disableChat?: boolean;
+      disableInvite?: boolean;
     },
   ): RoomInfoContextInterface['data'];
   microphoneDevice: (deviceId: deviceId) => void;
